GH-2925 move integration tests to their own github action and turn the slow shacl tests into integration tests
