+++ .github/workflows/pr-integration-tests.yml
@@ -1,4 +1,4 @@
-name: pull request verify
+name: pull request integration tests
 
 on: pull_request
 
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        jdk: [1.8, 15]
+        jdk: [1.8]
     steps:
     - uses: actions/checkout@v1
     - name: Set up JDK
@@ -21,12 +21,10 @@ jobs:
         key: ${{ runner.os }}-jdk${{ matrix.jdk }}-maven-${{ hashFiles('**/pom.xml') }}
         restore-keys: |
           ${{ runner.os }}-jdk${{ matrix.jdk }}-maven-
-    - name: Verify formatting
-      run: mvn -B formatter:validate impsort:check xml-format:xml-check
     - name: Build
       run: mvn -B clean install -Pquick,\!formatting
     - name: Run tests
-      run: mvn -B verify --file pom.xml
+      run: mvn -B integration-test -Pintegration -Dskip.surefire.tests
     - name: Print test failures
       if: failure()
       run: ./scripts/printTestResults.sh
