+++ core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/MultithreadedMemoryStoreIT.java
@@ -10,7 +10,7 @@ package org.eclipse.rdf4j.sail.shacl;
 import org.eclipse.rdf4j.sail.NotifyingSail;
 import org.eclipse.rdf4j.sail.memory.MemoryStore;
 
-public class MultithreadedMemoryStoreTest extends MultithreadedTest {
+public class MultithreadedMemoryStoreIT extends MultithreadedTest {
 	@Override
 	NotifyingSail getBaseSail() {
 		return new MemoryStore();
