+++ core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/TransactionalIsolationTest.java
@@ -555,283 +555,6 @@ public class TransactionalIsolationTest {
 	}
 
 	@Test
-	public void testIsolation2_multithreaded_READ_COMMITTED() throws Throwable {
-
-		for (int i = 0; i < 1000; i++) {
-
-			ShaclSail shaclSail = new ShaclSail(new MemoryStore());
-
-			SailRepository sailRepository = new SailRepository(shaclSail);
-			sailRepository.init();
-
-			try {
-				CountDownLatch countDownLatch = new CountDownLatch(2);
-
-				Runnable t1 = () -> {
-
-					try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-
-						connection.begin(IsolationLevels.READ_COMMITTED);
-						for (int k = 0; k < 1000; k++) {
-							addInTransaction(connection, "ex:steve" + k + " a ex:Person .");
-
-						}
-
-						countDownLatch.countDown();
-						try {
-							countDownLatch.await();
-						} catch (InterruptedException e) {
-							throw new IllegalStateException();
-						}
-
-						try {
-							connection.commit();
-						} catch (Throwable ignored) {
-						}
-
-					}
-
-				};
-
-				Thread thread1 = new Thread(t1);
-
-				thread1.start();
-
-				Runnable t2 = () -> {
-					try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-
-						connection.begin(IsolationLevels.READ_COMMITTED);
-						StringReader shaclRules = new StringReader(String.join("\n", "",
-								"@prefix ex: <http://example.com/ns#> .",
-								"@prefix sh: <http://www.w3.org/ns/shacl#> .",
-								"@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .",
-								"@prefix foaf: <http://xmlns.com/foaf/0.1/>.",
-
-								"ex:PersonShape",
-								"        a sh:NodeShape  ;",
-								"        sh:targetClass ex:Person ;",
-								"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										" ."));
-
-						try {
-							connection.add(shaclRules, "", RDFFormat.TURTLE, RDF4J.SHACL_SHAPE_GRAPH);
-						} catch (IOException e) {
-							throw new IllegalStateException();
-						}
-
-						countDownLatch.countDown();
-						try {
-							countDownLatch.await();
-						} catch (InterruptedException e) {
-							throw new IllegalStateException();
-						}
-
-						try {
-							connection.commit();
-						} catch (Throwable ignored) {
-						}
-					}
-
-				};
-
-				Thread thread2 = new Thread(t2);
-
-				thread2.start();
-
-				thread1.join();
-				thread2.join();
-
-				try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-					connection.begin();
-					ValidationReport validationReport = ((ShaclSailConnection) connection.getSailConnection())
-							.revalidate();
-
-					if (!validationReport.conforms()) {
-						Model statements = validationReport.asModel();
-						WriterConfig writerConfig = new WriterConfig();
-						writerConfig.set(BasicWriterSettings.PRETTY_PRINT, true);
-						writerConfig.set(BasicWriterSettings.INLINE_BLANK_NODES, true);
-						Rio.write(statements, System.out, RDFFormat.TURTLE, writerConfig);
-					}
-
-					assertTrue(validationReport.conforms());
-
-					connection.commit();
-				}
-			} finally {
-				sailRepository.shutDown();
-			}
-
-		}
-	}
-
-	@Test
-	public void testIsolation2_multithreaded_SNAPSHOT() throws Throwable {
-
-		for (int i = 0; i < 1000; i++) {
-
-			ShaclSail shaclSail = new ShaclSail(new MemoryStore());
-
-			SailRepository sailRepository = new SailRepository(shaclSail);
-			sailRepository.init();
-
-			try {
-
-				CountDownLatch countDownLatch = new CountDownLatch(2);
-
-				Runnable t1 = () -> {
-
-					try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-
-						connection.begin(IsolationLevels.SNAPSHOT);
-						for (int k = 0; k < 1000; k++) {
-							addInTransaction(connection, "ex:steve" + k + " a ex:Person .");
-
-						}
-
-						countDownLatch.countDown();
-						try {
-							countDownLatch.await();
-						} catch (InterruptedException e) {
-							throw new IllegalStateException();
-						}
-
-						try {
-							connection.commit();
-						} catch (Throwable ignored) {
-						}
-
-					}
-
-				};
-
-				Thread thread1 = new Thread(t1);
-
-				thread1.start();
-
-				Runnable t2 = () -> {
-					try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-
-						connection.begin(IsolationLevels.SNAPSHOT);
-						StringReader shaclRules = new StringReader(String.join("\n", "",
-								"@prefix ex: <http://example.com/ns#> .",
-								"@prefix sh: <http://www.w3.org/ns/shacl#> .",
-								"@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .",
-								"@prefix foaf: <http://xmlns.com/foaf/0.1/>.",
-
-								"ex:PersonShape",
-								"        a sh:NodeShape  ;",
-								"        sh:targetClass ex:Person ;",
-								"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										"        sh:property [",
-								"                sh:path ex:age ;",
-								"                sh:minCount 1 ;",
-								"        ] ;" +
-										" ."));
-
-						try {
-							connection.add(shaclRules, "", RDFFormat.TURTLE, RDF4J.SHACL_SHAPE_GRAPH);
-						} catch (IOException e) {
-							throw new IllegalStateException();
-						}
-
-						countDownLatch.countDown();
-						try {
-							countDownLatch.await();
-						} catch (InterruptedException e) {
-							throw new IllegalStateException();
-						}
-
-						try {
-							connection.commit();
-						} catch (Throwable ignored) {
-						}
-					}
-
-				};
-
-				Thread thread2 = new Thread(t2);
-
-				thread2.start();
-
-				thread1.join();
-				thread2.join();
-
-				try (SailRepositoryConnection connection = sailRepository.getConnection()) {
-					connection.begin();
-					ValidationReport validationReport = ((ShaclSailConnection) connection.getSailConnection())
-							.revalidate();
-
-					if (!validationReport.conforms()) {
-						Model statements = validationReport.asModel();
-						WriterConfig writerConfig = new WriterConfig();
-						writerConfig.set(BasicWriterSettings.PRETTY_PRINT, true);
-						writerConfig.set(BasicWriterSettings.INLINE_BLANK_NODES, true);
-						Rio.write(statements, System.out, RDFFormat.TURTLE, writerConfig);
-					}
-
-					assertTrue(validationReport.conforms());
-
-					connection.commit();
-				}
-			} finally {
-				sailRepository.shutDown();
-			}
-
-		}
-	}
-
-	@Test
 	public void testIsolation3() throws Throwable {
 		ShaclSail shaclSail = new ShaclSail(new MemoryStore());
