+++ core/sail/shacl/src/test/java/org/eclipse/rdf4j/sail/shacl/benchmark/BenchmarkIT.java
@@ -17,12 +17,12 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 import org.openjdk.jmh.runner.options.TimeValue;
 
 /**
- * This runs (almost) all the benchmarks during unit testing so that we know that we haven't broken any of the
+ * This runs (almost) all the benchmarks during integration testing so that we know that we haven't broken any of the
  * benchmarks.
  *
  * @author Håvard Ottestad
  */
-public class BenchmarkJunitTests {
+public class BenchmarkIT {
 
 	{
 		GlobalValidationExecutionLogging.loggingEnabled = false;
